rename/#0: config 패키지 이동
