+++ product/src/main/java/takeoff/logistics_service/msa/product/product/infrastructure/kafka/config/ProductTopicConfig.java
@@ -1,4 +1,4 @@
-package takeoff.logistics_service.msa.product.product.infrastructure.kafka;
+package takeoff.logistics_service.msa.product.product.infrastructure.kafka.config;
 
 import java.util.Map;
 import org.apache.kafka.clients.admin.NewTopic;
